backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
python-decouple>=3.8
mongoengine>=0.27
pymongo>=4.6
dnspython>=2.4
PyJWT>=2.8
bcrypt>=4.0
Pillow>=10.0
openpyxl>=3.1
